Fairness Algorithms

MLJFair provides with various algorithms that can help in mitigating bias and improving fairness metrics.

Introduction

These algorithms are wrappers. As demonstrated in last section, these wrappers can be used to compose a complex pipeline with more than 1 fairness algorithm. These wrappers can be used only with binary classifiers. These fairness algorithms have been divided into 3 categories based on the parts in the pipeline that the algorithm can control. These 3 categories are Preprocessing, Postprocessing and Inprocessing[WIP].

Preprocessing Algorithms

These are the algorithms that have control over the training data to be fed into machine learning model. This class of algorithms improves the representation of groups in the training data.

ReweighingSampling Algorithm

MLJFair.ReweighingSamplingWrapper — Type

ReweighingSamplingWrapper

ReweighingSamplingWrapper is a preprocessing algorithm wrapper in which Weights for each group-label combination is calculated. Using the calculated weights, rows are sampled uniformly. The weight is used to sample uniformly. The number of datapoints used to train after sampling from the reweighed dataset can be controlled by factor.

Missing docstring.

Missing docstring for MLJFair.ReweighingSamplingWrapper(::MLJBase.Model). Check Documenter's build log for details.

Reweighing Algorithm

This model being wrapped with this wrapper needs to support weights. If the model doesn't support training using weights, then error is thrown. In case weights are not supported by your desired model, them switch to ReweighingSampling Algorithm. To find the models in MLJ that support weights, execute:

using MLJ
models(x-> x.supports_weights)

MLJFair.ReweighingWrapper — Type

ReweighingWrapper

ReweighingWrapper is a preprocessing algorithm wrapper in which Weights for each group-label combination is calculated. These calculated weights are then passed to the classifier model which further uses it to make training fair.

Missing docstring.

Missing docstring for MLJFair.ReweighingWrapper(::MLJBase.Model). Check Documenter's build log for details.

MLJFair._calculateWeights — Function

Helper function for ReweighingWrapper and ReweighingSamplingWrapper. grps is an array of values of protected attribute. y is an array of ground truth values. Array of (Frequency) weights are returned from this function.

Postprocessing

These are the algorithms that have control over the final predictions. They can tweak final predictions to optimise fairness constraints.

Equalized Odds Algorithm

MLJFair.EqOddsWrapper — Type

EqOddsWrapper

It is a postprocessing algorithm which uses Linear Programming to optimise the constraints for Equalized Odds.

Missing docstring.

Missing docstring for MLJFair.EqOddsWrapper(::MLJBase.Model). Check Documenter's build log for details.

LinProg Algorithm

This algorithm supports all the metrics provided by MLJFair.

MLJFair.LinProgWrapper — Type

LinProgWrapper

It is a postprocessing algorithm that uses JuMP and Ipopt library to minimise error and satisfy the equality of specified specified measures for all groups at the same time. Automatic differentiation and gradient based optimisation is used to find probabilities with which the predictions are changed for each group.

Missing docstring.

Missing docstring for MLJFair.LinProgWrapper(::MLJBase.Model). Check Documenter's build log for details.

Composability

MLJFair provides you the ability to easily use multiple fairness algorithms on top of each other. A fairness algorithm can be added over another fairness algorithm by simply wrapping the previous wrapped model with the new wrapper. MLJFair handles everything else for you! The use of wrappers provides you the ability to add as many algorithms as you want!!

julia> using MLJFair, MLJ

julia> X, y, _ = @load_toydata;

julia> model = ConstantClassifier();

julia> wrappedModel = ReweighingSamplingWrapper(model; grp=:Sex);
ERROR: MethodError: no method matching ReweighingSamplingWrapper(::ConstantClassifier; grp=:Sex)
Closest candidates are:
  ReweighingSamplingWrapper(; classifier, grp, factor, rng) at /home/travis/build/ashryaagr/MLJFair.jl/src/algorithms/preprocessing/reweighing.jl:101
  ReweighingSamplingWrapper(!Matched::Symbol, !Matched::M, !Matched::Float64, !Matched::Union{Int64, Random.AbstractRNG}) where M<:Model at /home/travis/build/ashryaagr/MLJFair.jl/src/algorithms/preprocessing/reweighing.jl:86 got unsupported keyword argument "grp"

julia> wrappedModel2 = EqOddsWrapper(wrappedModel; grp=:Sex);
ERROR: UndefVarError: wrappedModel not defined

julia> mch = machine(wrappedModel2, X, y);
ERROR: UndefVarError: wrappedModel2 not defined

julia> fit!(mch)
ERROR: UndefVarError: mch not defined

julia> ŷ = predict(mch, X);
ERROR: UndefVarError: mch not defined