Fairness Datasets

To make it easy to try algorithms and metrics on various datasets, MLJFair shall be providing with the popular fairness datasets.

These datasets can be easily accesses using macros.

Following datasets shall be available in future through macros.

COMPAS Dataset

Adult Dataset

German Dataset

Bank Dataset

But currently, only a toy dataset is available.

Toy Data

This is a 10 row dataset that was used by authors of Reweighing Algorithm.

MLJFair.@load_toydata — Macro

Macro to read csv file of job data and convert columns to categorical. Returns the tuple (X, y, ŷ)

MLJFair.@load_toyfairtensor — Macro

Macro to create fairness Tensor for jobs.csv The fairness tensor will be created on the basis of the column Job Type. This column has 3 different values for job types.

julia> using MLJFair

julia> X, y, ŷ = @load_toydata;

julia> X
10×4 DataFrame
│ Row │ Sex  │ Ethnicity │ Highest degree │ Job type   │
│     │ Cat… │ Cat…      │ Categorical…   │ Cat…       │
├─────┼──────┼───────────┼────────────────┼────────────┤
│ 1   │ M    │ Native    │ H. school      │ Board      │
│ 2   │ M    │ Native    │ Univ.          │ Board      │
│ 3   │ M    │ Native    │ H. school      │ Board      │
│ 4   │ M    │ Non-nat.  │ H. school      │ Healthcare │
│ 5   │ M    │ Non-nat.  │ Univ.          │ Healthcare │
│ 6   │ F    │ Non-nat.  │ Univ.          │ Education  │
│ 7   │ F    │ Native    │ H. school      │ Education  │
│ 8   │ F    │ Native    │ None           │ Healthcare │
│ 9   │ F    │ Non-nat.  │ Univ.          │ Education  │
│ 10  │ F    │ Native    │ H. school      │ Board      │

julia> y
10-element CategoricalArrays.CategoricalArray{Int64,1,UInt32}:
 1
 1
 1
 1
 0
 0
 0
 1
 0
 1

julia> ŷ
10-element CategoricalArrays.CategoricalArray{Int64,1,UInt32}:
 1
 0
 1
 0
 1
 0
 1
 0
 1
 0

julia> ft = @load_toyfairtensor
MLJFair.FairTensor{3}([2 2; 0 0; 0 2]

[0 0; 2 1; 1 0], ["Board", "Education", "Healthcare"])

Below is the helper function that can be used to load custom dataset as well

MLJFair.load_dataset — Function

load_dataset(fpath, coercions)

Load one of standard dataset like Boston etc assuming the file is a comma separated file with a header.