Measures

The Measures have been divided into calcMetrics and boolmetrics depending on whether the metric returns Numerical value or Boolean value respectively.

MLJFair._ftIdx is a utility function that has been used to calculate metrics and shall be helpful while using MLJFair to inspect Fairness tensor values.

MLJFair._ftIdx — Function

_ftIdx(ft, grp)

Finds the index of grp (string) in ft.labels which corresponds to ft.mat. For Index i for the grp returned by this function ft[i, :, :] returns the 2D array [[TP, FP], [FN, TN]] for that group.

CalcMetrics

These metrics return a Numerical Value.

MLJFair.disparity — Function

disparity(M, ft; refGrp=nothing)

Computes disparity for fairness tensor ft with respect to an array of metrics M.

For any class A and a reference Group B, disparity = metric(A)/metric(B)

A dataframe is returned with disparity values for all combinations of metrics and classes. It contains a column named labels for the classes and has a column for disparity of each metric in M. The column names are metric names appended with _disparity.

Keywords

refGroup=nothing : The reference group

Please note that division by 0 will result in NaN

BoolMetrics

These metrics return a boolean value.

DemographicParity