Fairness Algorithms

MLJFair provides with various algorithms that can help in mitigating bias and improving fairness metrics.

Introduction

These algorithms are wrappers. As demonstrated in last section, these wrappers can be used to compose a complex pipeline with more than 1 fairness algorithm. These fairness algorithms have been divided into 3 categories based on the parts in the pipeline that the algorithm can control. These 3 categories are Preprocessing, Postprocessing and Inprocessing[WIP].

Preprocessing Algorithms

These are the algorithms that have control over the training data to be fed into machine learning model. This class of algorithms improves the representation of groups in the training data.

ReweighingSampling Algorithm

This algorithm also supports training data where more than 2 values are possible for sensitive attribute.

MLJFair.ReweighingSamplingWrapper — Type

ReweighingSamplingWrapper

ReweighingSamplingWrapper is a preprocessing algorithm wrapper in which Weights for each group-label combination is calculated. Using the calculated weights, rows are sampled uniformly. The weight is used to sample uniformly. The number of datapoints used to train after sampling from the reweighed dataset can be specified by noSamples.

MLJFair.ReweighingSamplingWrapper — Method

ReweighingSamplingWrapper(classifier; grp=:class, noSamples=-1)

Instantiates a ReweighingSamplingWrapper which wrapper the classifier with the Reweighing fairness algorithm together with sampling. The sensitive attribute can be specified by the parameter grp. noSamples indicates the number of datapoints used to train after sampling from the reweighed dataset. A negative or no value value for noSamples parameter instructs the algorithm to use the same number of datapoints as in original sample.

Reweighing Algorithm

This algorithm is currently not functioning expectedly. You can switch to ReweighingSamplingWrapper which provides similar functionality.

MLJFair.ReweighingWrapper — Type

ReweighingWrapper

ReweighingWrapper is a preprocessing algorithm wrapper in which Weights for each group-label combination is calculated. These calculated weights are then passed to the classifier model which further uses it to make training fair.

MLJFair.ReweighingWrapper — Method

ReweighingWrapper(classifier; grp=:class)

Instantiates a ReweighingWrapper which wrapper the classifier with the Reweighing fairness algorithm. The sensitive attribute can be specified by the parameter grp. If classifier doesn't support weights while training, an error is thrown.

MLJFair._calculateWeights — Function

Helper function for ReweighingWrapper and ReweighingSamplingWrapper. grps is an array of values of protected attribute. y is an array of ground truth values. Array of (Frequency) weights are returned from this function.

Postprocessing

These are the algorithms that have control over the final predictions. They can tweak final predictions to optimise fairness constraints.

Equalized Odds Algorithm

This algorithm currently supports training data with binary sensitive attribute.

MLJFair.EqOddsWrapper — Type

EqOddsWrapper

It is a postprocessing algorithm which uses Linear Programming to optimise the constraints for Equalized Odds.

MLJFair.EqOddsWrapper — Method

EqOddsWrapper

Instantiates EqOddsWrapper which wraps the classifier

Composability

MLJFair provides you the ability to easily use multiple fairness algorithms on top of each other. A fairness algorithm can be added over another fairness algorithm by simply wrapping the previous wrapped model with the new wrapper. MLJFair handles everything else for you! The use of wrappers provides you the ability to add as many algorithms as you want!!

julia> model = ConstantClassifier();

julia> wrappedModel = ReweighingSamplingWrapper(model; grp=:Sex);

julia> wrappedModel2 = EqOddsWrapper(wrappedModel; grp=:Sex);

julia> mch = machine(wrappedModel2, X, y);

julia> fit!(mch)
[ Info: Training [34mMachine{EqOddsWrapper} @ 6…13[39m.
[ Info: Training [34mMachine{ReweighingSamplingWrapper} @ 8…13[39m.
[34mMachine{EqOddsWrapper} @ 6…13[39m

julia> ŷ = predict(mch, X);