Fairness Datasets

To make it easy to try algorithms and metrics on various datasets, Fairness.jl provides you with the popular fairness datasets.

These datasets can be easily accesses using macros.

COMPAS Dataset

Fairness.@load_compas — Macro

Macro to load COMPAS dataset It is a reduced version of COMPAS Datset with 8 features and 6907 rows. The protected attributes are sex and race. The available features are used to predict whether a criminal defendant will recidivate(reoffend).

Returns (X, y)

julia> using Fairness

julia> X, y = @load_compas;

Adult Dataset

Fairness.@load_adult — Macro

Macro to Load the Adult dataset It has 14 features and 32561 rows. The protected attributes are race and sex. This dataset is used to predict whether income exceeds 50K dollars per year.

Returns (X, y)

German Credit Dataset

Fairness.@load_german — Macro

Load the full version of German credit dataset. This dataset has 20 features and 1000 rows. The protected attributes are gender_status and age (>25 is priviledged) Using the 20 features, it classifies the credit decision to a person as good or bad credit risks.

Returns (X, y)

Inspecting Datasets

To see the columns in dataset, their types and scientific types, you can use schema from MLJ.

julia> using Fairness, MLJ

julia> X, y = @load_adult;

julia> schema(X)
┌────────────────┬─────────────────────────────────┬────────────────┐
│ _.names        │ _.types                         │ _.scitypes     │
├────────────────┼─────────────────────────────────┼────────────────┤
│ age            │ Float64                         │ Continuous     │
│ workclass      │ CategoricalValue{String,UInt32} │ Multiclass{9}  │
│ fnlwgt         │ Float64                         │ Continuous     │
│ education      │ CategoricalValue{String,UInt32} │ Multiclass{16} │
│ education_num  │ Float64                         │ Continuous     │
│ marital_status │ CategoricalValue{String,UInt32} │ Multiclass{7}  │
│ occupation     │ CategoricalValue{String,UInt32} │ Multiclass{15} │
│ relationship   │ CategoricalValue{String,UInt32} │ Multiclass{6}  │
│ race           │ CategoricalValue{String,UInt32} │ Multiclass{5}  │
│ sex            │ CategoricalValue{String,UInt32} │ Multiclass{2}  │
│ capital_gain   │ Float64                         │ Continuous     │
│ capital_loss   │ Float64                         │ Continuous     │
│ hours_per_week │ Float64                         │ Continuous     │
│ native_country │ CategoricalValue{String,UInt32} │ Multiclass{42} │
└────────────────┴─────────────────────────────────┴────────────────┘
_.nrows = 32561

Toy Data

This is a 10 row dataset that was used by authors of Reweighing Algorithm. This dataset is intended to test ideas and evaluate metrics without calculating predictions. It is different from other macros as it returns (X, y, ŷ) instead of (X, y)

Fairness.@load_toydata — Macro

Macro to read csv file of job data (data/jobs.csv) and convert columns to categorical. Returns the tuple (X, y, ŷ)

Fairness.@load_toyfairtensor — Macro

Macro to create fairness Tensor for data/jobs.csv The fairness tensor will be created on the basis of the column Job Type. This column has 3 different values for job types.

julia> X, y, ŷ = @load_toydata;

julia> ft = @load_toyfairtensor
Fairness.FairTensor{3}([2 2; 0 0; 0 2]

[0 0; 2 1; 1 0], ["Board", "Education", "Healthcare"])

Other Datasets

You can try working with the vast range of datasets available through OpenML. Refer MLJ's OpenML documentation for the OpenML API. The id to be passed to OpenML.load can be found through OpenML site

julia> using MLJBase, Fairness

julia> using DataFrames

julia> data = OpenML.load(1480); # load Indian Liver Patient Dataset

julia> df = DataFrame(data) ;

julia> y, X = unpack(df, ==(:Class), name->true); # Unpack the data into features and target

julia> y = coerce(y, Multiclass); # Specifies that the target y is of type Multiclass. It is othewise a string.

julia> coerce!(X, :V2 => Multiclass, Count => Continuous); # Specifying which columns are Multiclass in nature. Converting from Count to Continuous enables use of more models.

Helper Functions

Fairness.ensure_download — Function

Checks whether the dataset is already present in data directory. Downloads it if not present.