Fairness Datasets

To make it easy to try algorithms and metrics on various datasets, MLJFair shall be providing with the popular fairness datasets.

These datasets can be easily accesses using macros.

German Credit Dataset

MLJFair.@load_german — Macro

Load the full version of German credit dataset. This dataset has 20 features and 1000 rows. The protected attributes are gender_status and age (>25 is priviledged) Using the 20 features, it classifies the credit decision to a person as good or bad credit risks.

Returns (X, y)

julia> using MLJFair

julia> X, y = @load_german;

Toy Data

This is a 10 row dataset that was used by authors of Reweighing Algorithm.

MLJFair.@load_toydata — Macro

Macro to read csv file of job data and convert columns to categorical. Returns the tuple (X, y, ŷ)

MLJFair.@load_toyfairtensor — Macro

Macro to create fairness Tensor for jobs.csv The fairness tensor will be created on the basis of the column Job Type. This column has 3 different values for job types.

julia> X, y, ŷ = @load_toydata;

julia> ft = @load_toyfairtensor
MLJFair.FairTensor{3}([2 2; 0 0; 0 2]

[0 0; 2 1; 1 0], ["Board", "Education", "Healthcare"])

Other Datasets

You can try working with the vast range of datasets available through OpenML. Refer MLJ's OpenML documentation for the OpenML API. The id to be passed to OpenML.load can be found through OpenML site

julia> using MLJBase, MLJFair;

julia> using DataFrames

julia> data = OpenML.load(1480); # load Indian Liver Patient Dataset

julia> df = DataFrame(data) ;

julia> y, X = unpack(df, ==(:Class), name->true); # Unpack the data into features and target

julia> y = coerce(y, Multiclass); # Specifies that the target y is of type Multiclass. It is othewise a string.

julia> coerce!(X, :V2 => Multiclass, Count => Continuous); # Specifying which columns are Multiclass in nature. Converting from Count to Continuous enables use of more models.

If you notice, the target y is either 1 or 2. But MLJFair supports only binary attributes. This shall be changed in future and we will support any 2 values which will be compared on the basis of levels. So, to proceed further, you need to map the values to binary.

y = map(y) do η
    η == "1" ? true : false
end;

Helper Functions

MLJFair.ensure_download — Function

Checks whether the dataset is already present in data directory. Downloads it if not present.